chore: some refactors
