+++ packages/styled-components/src/hooks/useTheme.js
@@ -1,0 +1,10 @@
+// @flow
+import { useContext } from 'react';
+import { ThemeContext } from '../models/ThemeProvider';
+
+const useTheme = () => {
+  const outerTheme = useContext(ThemeContext);
+  return outerTheme;
+}
+
+export default useTheme;
